requirements.txt
alembic==1.13.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
asn1crypto==1.5.1
bcrypt==4.0.1
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.8.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.115.0
Flask==2.3.2
Flask-Compress==1.13
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-secret-manager==2.24.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==1.26.4
openai==1.51.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdfkit==1.0.0
pg8000==1.31.5
Pillow==10.0.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydyf==0.11.0
PyJWT==2.9.0
PyMySQL==1.1.2
pyphen==0.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.13.0
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
pyzstd==0.18.0
qrcode==7.4.2
reportlab==4.2.5
requests==2.32.3
rsa==4.9.1
scramp==1.4.6
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.1
weasyprint==62.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.3.2
zopfli==0.2.3.post1
